fix(CartItems): Corrigindo excecoes no endpoint de delete quantidade de um look especifico

+++ app/controller/cart_items_controller.py
@@ -3,10 +3,10 @@ from db.database import get_db
 from sqlalchemy.ext.asyncio import AsyncSession
 from model.cart_items import CartItems
 from model.cart import Cart
-from model.look import Look
+from uuid import UUID
 from sqlalchemy.orm import joinedload
 from decimal import Decimal
-from schemas.cart_items import CartItemsAdd, CartItemsResponse, ListCartItemResponse, DeleteLookCartItemUser, CartItemsDelete
+from schemas.cart_items import *
 from sqlalchemy import select
 
 
@@ -20,7 +20,8 @@ router = APIRouter(prefix='/cart-items', tags=['Carts_Items'])
              summary="Adiciona um look no carrinho")
 async def add_look_cart(dados: CartItemsAdd, db: AsyncSession = Depends(get_db)):
 
-    look_in_cart = db.scalar(select(CartItems).where(CartItems.look_id == dados.look_id))
+    look_in_cart = await db.scalar(select(CartItems).where(CartItems.look_id == dados.look_id))
+
     if look_in_cart:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
                             detail = 'Look already exists in Cart of User')
@@ -62,8 +63,10 @@ async def get_looks_user(user_id: str, db: AsyncSession = Depends(get_db)):
     cart_items = result.all()
 
     price_total = Decimal("0.00")
+    
 
     for item in cart_items:
+        item.price_quantity_total = item.look.price * item.quantity
         if item.look.price:
             price_total += item.look.price
 
@@ -78,27 +81,55 @@ async def get_looks_user(user_id: str, db: AsyncSession = Depends(get_db)):
                summary="Exclui um look do carrinho de um usuario especifico")
 async def delete_look_cart_user(dados:DeleteLookCartItemUser, db: AsyncSession = Depends(get_db)):
     cart_user = await db.scalar(select(Cart).where(Cart.user_id == dados.user_id))
-    print(cart_user.id)
+    
     look = await db.scalar(select(CartItems).where(CartItems.look_id == dados.look_id and CartItems.cart_id == cart_user.id))
-    print(look)
+
     await db.delete(look)
     await db.commit()
 
 
-@router.put(path='/{look_id}/',
+@router.put(path='/remove/{look_id}/',
              status_code= status.HTTP_200_OK,
              response_model=CartItemsResponse,
-             summary="Diminui a quantidade de peças no carrinho")
-async def delete_one_look_cart(dados: CartItemsDelete, db: AsyncSession = Depends(get_db)):
-
+             summary="Diminui a quantidade de um look especifico no carrinho")
+async def delete_one_look_cart(look_id:UUID, dados: CartItemsDelete, db: AsyncSession = Depends(get_db)):
     cart_user = await db.scalar(select(Cart).where(Cart.id == dados.cart_id))
-    look = await db.scalar(select(CartItems).where(CartItems.cart_id == cart_user.id and CartItems.look_id == dados.look_id))
-    look.quantity -= 1
+    if cart_user == None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail = "Cart User not found")
+    look = await db.scalar(select(CartItems).where((CartItems.cart_id == cart_user.id) & (CartItems.look_id == look_id)))
+
+    if look == None:
+        raise HTTPException(status_code= status.HTTP_404_NOT_FOUND,
+                            detail = "Look not found in Cart of user")
+
+
+    if dados.quantity > look.quantity:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
+                            detail="The quantity sent for removal cannot be greater than the current quantity.")
+
+    look.quantity -= dados.quantity
+
     if look.quantity == 0:
         await db.delete(look)
         await db.commit()
         raise HTTPException(
-        status_code=400,
-        detail="Quantity item look is zero"
-    )
+        status_code=status.HTTP_200_OK,
+        detail="The quantity of the item look is equal to zero. We removed it from the cart."
+        )
+    
+    await db.commit()
+    return look
+
+
+@router.put(path='/add/{look_id}/',
+             status_code= status.HTTP_200_OK,
+             response_model=CartItemsResponse,
+             summary="Aumenta a quantidade de um look especifico no carrinho")
+async def add_one_more_look_cart(look_id:UUID, dados: CartItemsAddQuantity, db: AsyncSession = Depends(get_db)):
+    cart_user = await db.scalar(select(Cart).where(Cart.id == dados.cart_id))
+    look = await db.scalar(select(CartItems).where((CartItems.cart_id == cart_user.id) & (CartItems.look_id == look_id)))
+    look.quantity += dados.quantity
+    await db.commit()
+    await db.refresh(look)
     return look
